Update app.dart
Bumping the @platform packages to the newest. This will fix an issue with flutter_keychain (It's not used anymore)
there was a migration to the biometric_storage package... but the at_app cli was not referencing the newer versions of @platfrom packages
when creating a new app which leads to issues.